Iron ore restructure.
Rename it to "lode" so that its relationship with the more familiar
iron element is less direct, to maintain a higher baseline level of
suspension of disbelief.

Also restructure the way ores work in general, especially with
respect to normal stone.  Natural ore deposits are stone with a
texture overlaid.  Cobbled ore is the same ore "stain" texture with
alpha stripped and cobbles drawn over it.  Normal cobble is now
basically what "gravel ore" would look like.  This should allow
us to get a more consistent ore appearance with fewer textues.