Don't define "up-level dig times".
For now this is a stopgap to make sure that things aren't diggable
now that we want not to be.

This whole group level thing is kind of a mess.  We will probably
have to switch to a strategy of using fewer levels and more
different groups in the long run.  Based on the feedback I've been
getting in the MT Discord, it sounds like the number of tiers may
not be very flexible.