Fix itemstack repose/falling crash.
With item_eject trying to place stack nodes now instead of ents,
converting stack nodes to ents was causing an infinite loop.
Impart a tiny velocity on ents to make them ineligible for
automatic immediate stack placement.

Hacky.  Should probably have a separate API to indicate ejection
as an ent or as ent/stack or something.