Attempt to work around HUD issues by using image HUD.
On android (and possibly other very high DPI displays) the position
of the HUD is calculated incorrectly for text HUDs.  According to
rubenwardy, "the HUD sucks", so it seems unlikely that this will
get fixed soon.  Switch to using a sequence of image HUDs and a
bitmap font to render the hotbar text.  This is currently limited,
but expandable.