Make pickup and rearrange logic consistent again.
The new change preserves uncombined stacks when digging.  For
example, if you have dirt on the far right, tools on the far left,
and space between, and you dig dirt with the tools on the left,
it will create a NEW stack in the space between BEFORE it reaches
the dirt on the far right.

This behavior is necessary to ensure we can keep stacks separate
that we've separated for a specific purpose.

We also can't make the behavior vary based on whether using a tool
or not, because this would be even MORE jarring.