Initial proof-of-concept writing system.
It's barely usable/playable.

I'm not happy with how disjoint it is from the existing crafting
system, for little gain; there's a lot of on_nodepunch logic that
was added for this.

It also doesn't play very well.  Major issues:
- Awkward collision with existing logic, e.g. taking things from
  shelves, then drawing on side of shelf.  Probably need a group
  specifically for writable sufaces.
- Single action to rotate through all glyphs is bad.  Probably
  want one to change through the 8 shapes, and another to change
  rotation/reflection.
- Erasure is awkward, inconsistent, and not necessarily reliable.
  Should make it possible to recover charcoal in most cases, and
  use a unified set of hooks for it...

Should glyphs be pointable and run logic directly on node, or
should we keep it point-through...?