Add crude glass.
This makes glassmaking more challenging, as glass that is quenched
when not molded, or left wandering too long, will tend to quench to
crude glass, which is inferior.

This also prevents molten glass from wandering the countryside
forever from failed glassmaking attempts, and precludes keeping
molten glass in "captivity" as a crude infinite light source.