Add stone doors, tweak descriptions.
The alternate_loose/dig_in_place logic for cobble is interfering
with how stone doors work when dug, but it's still playable and
lossless, even though it's a bit tedious.  I'd like to understand
why it's happening this way and I can't seem to bypass it easily,
but it may be justifiable to leave it this way.