Squeeze sponges to produce temporary splashes of water.
Pummel a sponge and it will release water if there is space
nearby, though it will be reabsorbed by the sponge in a few
seconds.

This CAN be used to wet other sponges (i.e. to have a managed
supply of water if working far from oceans) but CANNOT be used
to create infinite springs.

If the original sponge is removed (and there is no other sponge
to absorb the water) then the water will be removed as well.