Rebrand as "puzzle/adventure".
This game is pretty hard to classify.  In principle it's sort of
like those old classic point-and-click "adventure" games where you
wander an open world, getting things and trying to use things on
other things, until you find combinations that do something
interesting and useful.

The game is mostly focused on "exploration" but not necessarily the
"traveling long distances over land" kind, but often the "exploring
the working principles of things" kind.