Support MagikEh's Pez Dispensers.
When removing a stack, if there is a stack above, move it
immediately into place of the removed stack.  This works
recursively, so stack nodes never have a chance to break out
into item entities, fall to the bottom and make a huge mess.

Unfortunately there isn't a very good visual for the stack
shifting down yet, but at least the gameplay is there.