Make players non-pointable for now.
Direct PvP is not (yet) a meaningful thing in NodeCore,
so there's no real benefit to targeting a player with a
tool.  However, having players be pointable CAN cause
issues with spectators blocking the selection path of
legit players in multiplayer.  This allows players to dig
around such spectators, and other players.

Players are not really "of the world" so this makes a
certain kind of sense...