New crack texture/animation.
- Now has 24 stages instead of 12.
- First stage is blank, to prevent very long digs from
  giving player too much positive feedback.
- Rotation effect.

The key thing here is that for the first 1/24 of the
dig animation, no positive feedback is given to the
player to indicate that the dig is "working."  This
should discourage new players from being distracted
by the emergency hand-digging feature.

More dig stages also help players get a better sense
of progress as they're digging things that do take a
long time, since digging in NodeCore tends to be
slower overall than digging in most MT games.

The rotation was thrown in as an afterthought, but
helps give a sense of progress.  Hopefully it's not
too distracting or seizure-inducing on very short dig
times; it ends up sort of all blurring into a circle to me
when digging with advanced tools...