Bake hotbar sel into hotbar.
This prevents the race between client-side changing
the hotbar selection index and server-side correcting
the selection image.

As a side effect, the hotbar slots no longer change
in size based on wield index, but this effect is not
very significant, and could be worked around if needed by adding padding to the normal
hotbar texture.

This branch depends on upstream PR #9271 being
merged for the MT client.