Don't pre-scatter ejected items.
Supposedly some stacks are clipping through the nodes
surrounding them when ejected; this is theoretically
possible if the stacks are full enough, based on the
X/Z offsets and the collisionbox sizes.

Always start items near the center of the ejecting node
position, to minimize (in theory, eliminate) this risk.