Fix bandolier slot visuals.
Alpha blending is apparently broken for
upright_sprite as of the latest v5.2-dev.

Use a simpler, more solid texture for
empty bandolier slots that looks a bit
more reminiscent of a stitched-on pocket
but still abstract.

Just hide selected slot altogether instead
of showing an alternate "selected slot"
texture.