Drop rod chisel as an item instead
Instead of instantly digging the rod, which makes it look like
it disappeared, try just "loosening" it into an item so it can be
scooped up without the pick.

It was always also possible to place the rod as an item and
pound that in, making it easier to pick up, but this may not
have been obvious to most players.