Default all sounds to ephemeral
NodeCore was already using sounds this way, i.e. assuming
they can't meaningfully be managed (may not be heard by
players just entering the area) so might as well get the
supposed performance benefit.

Added an override to allow sounds to be marked as NOT
ephemeral just in case it's needed.