Guard against future sprite prop noise
MT apparently internally ignores sprite properties on
entities that are not using a sprite visual, and returns
dummy values, rather than preserving what was sent
in, on get_properties().

Guard against set_property() noise on future/mod
calls explicitly rather than relying on the caller not
sending these props in when not applicable.