Revert player buoyancy
This feature is too annoying in practice and makes
the player harder to control when in water under most
circumstances, and isn't worth it for the cases it does
handle.

Buffing player breath should help alleviate drowning
problems due to network issues or chat.  Players may
just have to manually resurface more frequently.

This feature could be reconsidered for inclusion if a
better way to do it could be found, and if it could limit
gameplay impact.  Ideally the player should simply
remain at a fixed location when not moving (i.e. the
player character should counteract any drift
automatically) but that would require engine/client
changes.