Nerf lux tool durability
These things are way too durable as they are, and it's
actually kind of game-breaking.  Players with lux
tools are in an entirely different class from players
without.

Instead, players should be forced to invest more in
lux reactors, or rely more on tempered lode as an
"everyday" tool and use lux for special projects.

Boosted lux is exempt from this nerf, and is now 4
times as durable as ordinary lux.  It might not be a
bad idea to allow them not to experience any wear
at all when boosted...