Unalias hotbar slot textures
This was causing a texture from texture packs
to be reinterpreted differently, causing buggy-
looking hotbars, and TP authors would not
have been able to change the hotbar to work
with both versions.

Now before and after use different sets of
textures so a single TP can safely support
both before and after.