Alternative hotbar approach
We still change the background, but only depend
on the state of the slot, not on wield index.  Go
back to a traditional slot selection ring for the
selection index.

Slot backgrounds are a little bolder normally, and
when broken, especially bold around the broken
areas.

The selection ring itself is broken so that the
gaps in broken slots are still "open" to the outside
when the selection ring is in place.