"Dev Version" = "Dev/CDB Version"
ContentDB's .gitattributes export-subst support
has been broken for a while now, and apparently
won't be fixed (it's depending on an upstream
that is similarly ignoring the issue).

We need to account for the fact that what looks
like a dev version might actually be a CDB
version that's simply not tagged right.