Redo pickup interceptor using base metatables
The new method of intercepting item addition into
player inventories hooks into the base metatables for
the userdata types only once per type, instead of
trying to substitute every inventory object with a
table (which crashed some things) and does a lookup
to find the player per inv.  This should hopefully be
more reliable, and avoid some of the leaks to
rearrangement we had before.

If we can finally stop hitting the player inv rearrange
safety net then it can be removed.