Language bidirectional filling
Share strings across variants for a common
langauge, if we don't have a string for that
specific variant.  This should help us get better
overall translation coverage.

This is done as a "filter" on converting the raw
weblate data to MT locale files, so we don't get
the purity of the original translations ruined.