Finally retire old inventory rearrange logic
It seems as though the new intercept hooks are catching
everything, and since this logic was disabled hardly anyone
noticed.  The few rearrangements that were happening were
probably caused by inventory clear/restore mods unrelated to
NC operation.  This code was quite complex and removing it
will simplify the project.