Revert ABM time budget configuration.
The idea was to have Minetest spend more of its total available
time running ABMs so that ABMs are less likely to fall behind.
The problem is that Minetest's ABM time management is so completely
TERRIBLE that this causes huge lagspikes and makes the game
largely unplayable, so we're required to waste this time instead.