Falling nodes displace itemstacks
The realism theory is that item stacks are "looser"
and more easily pushed out of the way.  The
gameplay theory is that it's getting annoying
having machines like log or charcoal choppers
getting blocked by items under the falling nodes.

This is also hopefully part of falling-node stack
order reform, to fix that nodes sometimes fall
through each other.