Fix digging undiggables at high lag
Apparently (e.g. on Kimapr's servers) it's possible
to dig "undiggable" nodes under extreme lag.

Theory:

- Place something like cobble next to water source
- Dig with infused mattock through all stages

The client will predict a cobble->loose cobble
transition, then let you dig the loose cobble, and
then send the "dig" command to the server, but
on the server side the cobble instead was dug
directly, and then water flowed in to fill the space
where the client thought the loose cobble would
be, so the client digs the water, and you get
flowing water in your inventory.