Dump inventory on all teleports by default
A low-level player teleportation must be explicitly
declared nodecorian via a keepinv property on
the position, or the player must be exempted via
the keepinv priv, to be allowed to teleport without
leaving the inventory behind.

This means that naive mods like /home commands
will now work correctly and teleport only the
player, and not items.