Unprotect item stack nodes
This prevents a problem where players
can drop items into other players'
protected areas but cannot dig them.

Other solutions to prevent players from
placing items into protected areas in the
first place may be possible, and may still
be explored, but are not foolproof and so
this acts as a fallback at least.